LPD-54630 Isolate and simplify functions
